+++ app/Http/Actions/Owner/Product/ProductStoreAction.php
@@ -3,7 +3,7 @@
 namespace App\Http\Actions\Owner\Product;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Image\ImageStoreRequest;
+use App\Http\Requests\Owner\Image\ImageStoreRequest;
 use App\Models\Image;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\DB;
