+++ app/Http/Actions/Owner/Shop/ShopUpdateAction.php
@@ -3,7 +3,7 @@
 namespace App\Http\Actions\Owner\Shop;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Owner\ShopUpdateRequest;
+use App\Http\Requests\Owner\Shop\ShopUpdateRequest;
 use App\Models\Shop;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\DB;
