+++ app/Http/Requests/Owner/Image/ImageUpdateRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests\Image;
+namespace App\Http\Requests\Owner\Image;
 
 use Illuminate\Foundation\Http\FormRequest;
