+++ app/Http/Requests/Owner/Product/ProductUpdateRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests\Product;
+namespace App\Http\Requests\Owner\Product;
 
 use Illuminate\Foundation\Http\FormRequest;
